Added support for automatically repartitioning the container-vm images.
The container-vm image does not seem to play nicely with fdisk(8) because it
hangs after reboot when we modify the partition table. Instead, we install the
cloud-init package and use the "growroot" tool it contains to do this for us,
giving us a much simpler solution.

However, while it works well for Debian and derivatives (which container-vm is),
it does not seem to work as well with RHEL, CentOS, and SUSE, so we continue to
use the fdisk(8) + resize2fs(8) approach there.

Other changes:
* Added a test script (gcutil_test.sh) which lets us easily bring up a VM
  instance of each type of OS image we support, and query for free space using
  df(1) after the automatic repartitioning is complete.
* Renamed instances from fdisk-## to instance-## to account for the new method
* Collect PIDs for all processes started in the background and wait for them all
  in parallel, so we know when the entire operation has been completed.