Separate license comment from rest of file
Some editors are configured to fold the top-of-file comments for ease of reading
as they are typically boilerplate comments; by including meaningful comments in
the same block, they will also be hidden.

By separating the license comment from meaningful top-of-file comments, we
enable easily folding the licensing comment while still being able to see the
file-specific comment at the top while editing the file in such an editor.