Import Jupyter notebook cleanup tool
This imports the tooling I wrote previously for another of my projects:
https://github.com/mbrukman/reimplementing-ml-papers to be able to clean up and
simplify Jupyter notebooks which may be processed by a combination of Colab, VS
Code, etc. and make sure they don't add any extraneous fields or diffs to
notebooks, so that the diffs are easier to read, and the notebooks themselves
are minimal.